move WorkerInfo to utils
