+++ lmcache/v1/cache_controller/utils.py
@@ -1,0 +1,14 @@
+# SPDX-License-Identifier: Apache-2.0
+# Standard
+from dataclasses import dataclass
+
+
+@dataclass
+class WorkerInfo:
+    instance_id: str
+    worker_id: int
+    ip: str
+    port: int
+    distributed_url: str
+    registration_time: float
+    last_heartbeat_time: float
